We are suspecting the deletebyquery causes an issue due to multiple versions of the same object being handled in elasticsearch at the same time. For some reason, "id=x" is attempted to be deleted for "version=a" and "version=b", which fails. Added "conflict: proceed" to deletebyquery, as described in: https://www.elastic.co/guide/en/elasticsearch/reference/current/docs-delete-by-query.html . This should deleted _all_ matching documents and not stop at the first one that is matched or failed.

Related work items: #188304

+++ index/src/main/java/org/fao/geonet/index/es/EsRestClient.java
@@ -43,6 +43,7 @@ import co.elastic.clients.transport.rest_client.RestClientTransport;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.google.common.collect.Lists;
 import jakarta.json.spi.JsonProvider;
 import org.apache.commons.collections.MapUtils;
 import org.apache.commons.lang.StringUtils;
@@ -365,15 +366,15 @@ public class EsRestClient implements InitializingBean {
         }
 
         DeleteByQueryRequest request = DeleteByQueryRequest.of(
-            b -> b.index(new ArrayList<>(Arrays.asList(index)))
+            b -> b.index(Lists.newArrayList(index))
                 .q(query)
-                .refresh(true));
+                .refresh(true)
+                .conflicts(Conflicts.Proceed));
 
         final DeleteByQueryResponse deleteByQueryResponse =
             client.deleteByQuery(request);
 
-
-        if (deleteByQueryResponse.deleted() >= 0) {
+        if (deleteByQueryResponse.deleted() != null && deleteByQueryResponse.deleted() >= 0) {
             return String.format("Record removed. %s.", deleteByQueryResponse.deleted());
         } else {
             StringBuilder stringBuilder = new StringBuilder();
@@ -388,6 +389,7 @@ public class EsRestClient implements InitializingBean {
 
     /**
      * Get the complete document from the index.
+     *
      * @param id For record index, use UUID.
      * @return the source as Map
      */
